Improved user experience during registration - Added a "Confirm Password" input field to the registration form to allow users to verify their password - Implemented custom error handling to provide clear feedback when:   - The username or email is already in use   - The passwords do not match

+++ Frontend/authentication/register.js
@@ -1,50 +1,55 @@
 let submitRegisterForm = document.getElementById("form-register");
 let email;
+
 submitRegisterForm.addEventListener("submit", async function(event) {
     event.preventDefault();
 
     email = document.getElementById("email").value;
     let password = document.getElementById("password").value;
+    let confirmPassword = document.getElementById("confirm-password").value;
     let username = document.getElementById("username").value;
 
-    // Make POST request
-    try {
-        const response = await fetch(
-            "http://localhost:8080/api/authentication/register",
-            {
-                method: "POST",
-                headers: {
-                    "Content-Type": "application/json",
-                },
-                body: JSON.stringify({ email: email, password: password, username: username }),
-            }
-        );
+    if (password === confirmPassword) {
+        // Make POST request
+        try {
+            const response = await fetch(
+                "http://localhost:8080/api/authentication/register",
+                {
+                    method: "POST",
+                    headers: {
+                        "Content-Type": "application/json",
+                    },
+                    body: JSON.stringify({ email: email, password: password, username: username }),
+                }
+            );
 
-        if (response.ok) {
-           // Send user email in index.html
-           localStorage.setItem("userEmail", email);
-           // Redirect to success page if authentication is successful
-           window.location.href = "login.html"; 
-        } else {
-            showError();
+            if (response.ok) {
+                // Send user email in index.html
+                localStorage.setItem("userEmail", email);
+                // Redirect to success page if authentication is successful
+                window.location.href = "login.html"; 
+            } else {
+                const errorData = await response.text();
+                showError(errorData);
+            }
+        } catch (error) {
+            showError("Something went wrong!");
         }
-    } catch (error) {
-        showError();
+    } else {
+        showError("Passwords must be the same!");
     }
 });
-
 
 document.getElementById("btn-login").addEventListener("click", function(){
     window.location.href = "login.html";
 });
 
-
 let spanError = document.createElement('span');
 let divButtons = document.getElementById("div-buttons");
 
-function showError() {
+function showError(message) {
     spanError.classList.add('span-error'); 
-    spanError.innerText = "Something went wrong!"; 
+    spanError.innerText = message; 
     submitRegisterForm.insertBefore(spanError, divButtons); 
     setTimeout(function () {
         spanError.remove();
